Time: 2 ms (64.02%), Space: 7.6 MB (92.38%) - LeetHub

+++ 0063-unique-paths-ii/0063-unique-paths-ii.cpp
@@ -1,13 +1,12 @@
 class Solution {
 public:
-    long long dp[101][101];
-    // int mod = 1e9+7;
+    int dp[101][101];
     int solve(vector<vector<int>> &grid,int i,int j){
         int n = grid.size(), m = grid[0].size();
         if(i==-1 or j==-1 or grid[i][j]==1) return 0;
         if(i==0 and j==0) return 1;
         if(-1 != dp[i][j]) return dp[i][j];
-        long long ans = 0;
+        int ans = 0;
         ans += solve(grid,i,j-1);
         ans += solve(grid,i-1,j);
         return dp[i][j] = ans;
